% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{CDFPlot}
\alias{CDFPlot}
\alias{CDFPlot,FitDist-method}
\alias{CDFPlot,ANY-method}
\title{Plotting the CDF of data and fitted distribution}
\usage{
CDFPlot(object, ...)

\S4method{CDFPlot}{FitDist}(object, n = missing)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}

\item{n}{Number of samples, should not be used in current setting}
}
\description{
Plotting the CDF of data and fitted distribution
}
