% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{CDFPlot}
\alias{CDFPlot}
\title{Plotting the CDF of data and fitted distribution}
\usage{
CDFPlot(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Plotting the CDF of data and fitted distribution
}
