% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{ChiSqrTest}
\alias{ChiSqrTest}
\alias{ChiSqrTest,FitDist-method}
\alias{ChiSqrTest,ANY-method}
\title{Chi-Squared Test}
\usage{
ChiSqrTest(object, ...)

\S4method{ChiSqrTest}{FitDist}(object)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Chi-Squared Test
}
