% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{ChiSqrTest}
\alias{ChiSqrTest}
\title{Chi-Squared Test}
\usage{
ChiSqrTest(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Chi-Squared Test
}
