% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{claimFitting}
\alias{claimFitting}
\title{Claim data fitting analysis at line/type/status level}
\usage{
claimFitting(object, claimData, ...)
}
\arguments{
\item{object}{Simulation object}

\item{claimData}{claim data including existing claims for RBNER and claim reopenness analysis}

\item{startDate}{Date after which claims are analyzed}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR}

\item{lineList}{List of business lines to be included in claim fitting}

\item{typeList}{List of claim types to be included in claim fitting}

\item{discreteDist}{List of discrete distributions to try fitting (report lag, settlemet lag, frequency)}

\item{continuousDist}{List of continuous distribution to try fitting (severity)}

\item{copulaList}{List of copula to try fitting}

\item{fReportLag}{Boolean variable to indicate whether report lag needs to be fitted.}

\item{fSettlementLag}{Boolean variable to indicate whether settlement lag needs to be fitted.}

\item{fFrequency}{Boolean variable to indicate whether monthly frequency needs to be fitted.}

\item{fSeverity}{Boolean variable to indicate whether severity needs to be fitted.}

\item{fSSRCorrelation}{Boolean variable to indicate whether copula among severity, report lag and settlement lag needs to be fitted.}

\item{fFreqCorrelation}{Boolean variable to indicate whether copula among frequencies of business lines needs to be fitted.}

\item{copulaTest}{Whether to test copula. The testing could take a very long time.}

\item{iTotalLoss}{Boolean variable to indicate whether total loss before deductible and limit is available for severity fitting}

\item{fDeductible}{Boolean variable to indicate whether deductible empirical distribution needs to be fitted.}

\item{fLimit}{Boolean variable to indicate whether limit empirical distribution needs to be fitted.}

\item{check}{Boolean variable to indicate whether graph of each tried distribution fitting needs to be generated and saved.}
}
\description{
Claim data fitting analysis at line/type/status level
}
