% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Type.R
\name{claimSample}
\alias{claimSample}
\title{Claim simulation at line/type/status level}
\usage{
claimSample(object, ...)
}
\arguments{
\item{object}{ClaimType object}

\item{claimData}{claim data including existing claims for RBNER and claim reopenness analysis}

\item{startDate}{Date from which claim data is included in the analysis}

\item{evaluationDate}{Date of evaluation}
}
\description{
Claim simulation at line/type/status level
}
