% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\docType{methods}
\name{claimSimulation}
\alias{claimSimulation}
\alias{claimSimulation,Simulation-method}
\alias{claimSimulation,ANY-method}
\title{Claim simulation at line/type/status level}
\usage{
claimSimulation(object, ...)

\S4method{claimSimulation}{Simulation}(object, claimData = data.frame(),
  startDate = as.Date("2012-01-01"), evaluationDate = as.Date("2016-12-31"),
  futureDate = as.Date("2017-12-31"), append = TRUE)
}
\arguments{
\item{object}{Simulation object}

\item{...}{Additional parameters that may or may not be used.}

\item{claimData}{claim data including existing claims for RBNER and claim reopenness analysis;}

\item{startDate}{Date after which claims are analyzed;}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR;}

\item{futureDate}{Date of evaluation for UPR (future claims).}

\item{append}{Boolean variable to indicate whether existing simulation results need to be kept.}
}
\description{
Claim simulation at line/type/status level
}
\examples{
library(cascsim)
data(claimdata)
lines<-c("Auto","Property","Liab")
types<-c("N","H")
#exposure index
index1 <- new("Index",monthlyIndex=c(rep(1,11),cumprod(c(1,rep(1.5^(1/12),11))),
cumprod(c(1.5,rep((1.3/1.5)^(1/12),11))),
cumprod(c(1.3,rep((1.35/1.3)^(1/12),11))),cumprod(c(1.35,rep((1.4/1.35)^(1/12),11))),rep(1.4,301)))
#severity index
index2 <- new("Index",monthlyIndex=c(cumprod(c(1,rep(1.03^(1/12),59))),rep(1.03^(5),300)))
objan <- new("ClaimType", line="Auto",claimType="N",exposureIndex=index1,severityIndex=index2)
objah <- new("ClaimType", line="Auto",claimType="H",exposureIndex=index1,severityIndex=index2)
objpn <- new("ClaimType", line="Property",claimType="N",exposureIndex=index1,severityIndex=index2)
objph <- new("ClaimType", line="Property",claimType="H",exposureIndex=index1,severityIndex=index2)
objln <- new("ClaimType", line="Liab",claimType="N",exposureIndex=index1,severityIndex=index2)
objlh <- new("ClaimType", line="Liab",claimType="H",exposureIndex=index1,severityIndex=index2)
objlist <- list(objan,objah,objpn,objph,objln,objlh)
simobj <- new("Simulation",lines=lines,types=types,claimobjs=objlist,iFit=TRUE,
iCopula=TRUE, iReport=TRUE)
simobj <- claimFitting(simobj,claimdata)
library(cascsim)
data(claimdata)
lines<-c("Auto","Property","Liab")
types<-c("N","H")
#IBNR Only
library(cascsim)
data(claimdata)
lines<-c("Auto","Property","Liab")
types<-c("N","H")
objan <- new("ClaimType", line="Auto",claimType="N")
objah <- new("ClaimType", line="Auto",claimType="H")
objpn <- new("ClaimType", line="Property",claimType="N")
objph <- new("ClaimType", line="Property",claimType="H")
objln <- new("ClaimType", line="Liab",claimType="N")
objlh <- new("ClaimType", line="Liab",claimType="H")
objlist <- list(objan,objah,objpn,objph,objln,objlh)
simobj <- new("Simulation",lines=lines,types=types,iRBNER=FALSE,iROPEN=FALSE,iIBNR=TRUE,
iUPR=FALSE,claimobjs=objlist,simNo=2)
simdata <- claimSimulation(simobj)
simSummary <- simSummary(simobj,simdata)
simobj@iReport <- TRUE
simReport(simobj,simSummary)

#IBNR and UPR Only
#library(cascsim)
#data(claimdata)
#lines<-c("Auto","Property","Liab")
#types<-c("N","H")
#objan <- new("ClaimType", line="Auto",claimType="N")
#objah <- new("ClaimType", line="Auto",claimType="H")
#objpn <- new("ClaimType", line="Property",claimType="N")
#objph <- new("ClaimType", line="Property",claimType="H")
#objln <- new("ClaimType", line="Liab",claimType="N")
#objlh <- new("ClaimType", line="Liab",claimType="H")
#objlist <- list(objan,objah,objpn,objph,objln,objlh)
#simobj <- new("Simulation",lines=lines,types=types,iRBNER=FALSE,iROPEN=FALSE,iIBNR=TRUE,
#iUPR=TRUE,claimobjs=objlist,simNo=2)
#simdata <- claimSimulation(simobj)
#simSummary <- simSummary(simobj,simdata)
#simobj@iReport <- TRUE
#simReport(simobj,simSummary)

#All four claim classes: RBNER, ROPEN, IBNR and UPR
#library(cascsim)
#data(claimdata)
#lines<-c("Auto","Property","Liab")
#types<-c("N","H")
#objan <- new("ClaimType", line="Auto",claimType="N")
#objah <- new("ClaimType", line="Auto",claimType="H")
#objpn <- new("ClaimType", line="Property",claimType="N")
#objph <- new("ClaimType", line="Property",claimType="H")
#objln <- new("ClaimType", line="Liab",claimType="N")
#objlh <- new("ClaimType", line="Liab",claimType="H")
#objlist <- list(objan,objah,objpn,objph,objln,objlh)
#simobj <- new("Simulation",lines=lines,types=types,iRBNER=TRUE,iROPEN=TRUE,iIBNR=TRUE,
#iUPR=TRUE,claimobjs=objlist,simNo=2)
#simdata <- claimSimulation(simobj,claimData=claimdata)
#simSummary <- simSummary(simobj,simdata)
#simTriangle <- simTriangle(simobj,claimdata,simdata)
#simobj@iReport <- TRUE
#simReport(simobj,simSummary,simTriangle)
#multicore computing
#simobj@ncores<-4
#simobj@simNo<-2
#simdata <- claimSimulation(simobj,claimdata)
#simSummary <- simSummary(simobj,simdata)
#simTriangle <- simTriangle(simobj,claimdata,simdata)
#simobj@iReport <- TRUE
#simReport(simobj,simSummary)
}
