% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{claimSimulation}
\alias{claimSimulation}
\title{Claim simulation at line/type/status level}
\usage{
claimSimulation(object, ...)
}
\arguments{
\item{object}{Simulation object}

\item{claimData}{claim data including existing claims for RBNER and claim reopenness analysis}

\item{startDate}{Date after which claims are analyzed}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR}

\item{futureDate}{Date of evaluation for UPR (future claims)}
}
\description{
Claim simulation at line/type/status level
}
\examples{
library(cascsim)
data(claimdata)
lines<-c("Auto","Property","Liab")
types<-c("N","H")
#IBNR Only
library(cascsim)
data(claimdata)
lines<-c("Auto","Property","Liab")
types<-c("N","H")
objan <- new("ClaimType", line="Auto",claimType="N")
objah <- new("ClaimType", line="Auto",claimType="H")
objpn <- new("ClaimType", line="Property",claimType="N")
objph <- new("ClaimType", line="Property",claimType="H")
objln <- new("ClaimType", line="Liab",claimType="N")
objlh <- new("ClaimType", line="Liab",claimType="H")
objlist <- list(objan,objah,objpn,objph,objln,objlh)
simobj <- new("Simulation",lines=lines,types=types,iRBNER=FALSE,iROPEN=FALSE,iIBNR=TRUE,iUPR=FALSE,claimobjs=objlist,simNo=2)
simdata <- claimSimulation(simobj)
simSummary <- simSummary(simobj,simdata)
simobj@iReport <- TRUE
simReport(simobj,simSummary)

#IBNR and UPR Only
library(cascsim)
data(claimdata)
lines<-c("Auto","Property","Liab")
types<-c("N","H")
objan <- new("ClaimType", line="Auto",claimType="N")
objah <- new("ClaimType", line="Auto",claimType="H")
objpn <- new("ClaimType", line="Property",claimType="N")
objph <- new("ClaimType", line="Property",claimType="H")
objln <- new("ClaimType", line="Liab",claimType="N")
objlh <- new("ClaimType", line="Liab",claimType="H")
objlist <- list(objan,objah,objpn,objph,objln,objlh)
simobj <- new("Simulation",lines=lines,types=types,iRBNER=FALSE,iROPEN=FALSE,iIBNR=TRUE,iUPR=TRUE,claimobjs=objlist,simNo=2)
simdata <- claimSimulation(simobj)
simSummary <- simSummary(simobj,simdata)
simobj@iReport <- TRUE
simReport(simobj,simSummary)

#All four claim classes: RBNER, ROPEN, IBNR and UPR
library(cascsim)
data(claimdata)
lines<-c("Auto","Property","Liab")
types<-c("N","H")
objan <- new("ClaimType", line="Auto",claimType="N")
objah <- new("ClaimType", line="Auto",claimType="H")
objpn <- new("ClaimType", line="Property",claimType="N")
objph <- new("ClaimType", line="Property",claimType="H")
objln <- new("ClaimType", line="Liab",claimType="N")
objlh <- new("ClaimType", line="Liab",claimType="H")
objlist <- list(objan,objah,objpn,objph,objln,objlh)
simobj <- new("Simulation",lines=lines,types=types,iRBNER=TRUE,iROPEN=TRUE,iIBNR=TRUE,iUPR=TRUE,claimobjs=objlist,simNo=2)
simdata <- claimSimulation(simobj,claimData=claimdata)
simSummary <- simSummary(simobj,simdata)
simTriangle <- simTriangle(simobj,claimdata,simdata)
simobj@iReport <- TRUE
simReport(simobj,simSummary,simTriangle)
#multicore computing 
simobj@ncores<-4
simobj@simNo<-9
simdata <- claimSimulation(simobj,claimdata)
simSummary <- simSummary(simobj,simdata)
simTriangle <- simTriangle(simobj,claimdata,simdata)
simobj@iReport <- TRUE
simReport(simobj,simSummary)
}
