% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula.R
\docType{methods}
\name{copulaDataPlot}
\alias{copulaDataPlot}
\alias{copulaDataPlot,CopulaObj-method}
\alias{copulaDataPlot,ANY-method}
\title{Experience data plotting.}
\usage{
copulaDataPlot(object, ...)

\S4method{copulaDataPlot}{CopulaObj}(object)
}
\arguments{
\item{object}{Copula Object}

\item{...}{Additional parameters that may or may not be used}
}
\description{
Experience data plotting.
}
