% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{Density}
\alias{Density}
\title{Density function.}
\usage{
Density(object, x, ...)
}
\arguments{
\item{object}{Distribution Object}

\item{x}{Variable value}

\item{...}{Additional function arguments}
}
\description{
Density function.
}
\examples{
xPareto <- new("Pareto",p1=20,p2=3)
Density(xPareto,50)
}
