% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{Distribution-class}
\alias{Distribution-class}
\title{An S4 class to represent a distribution, either parametric or non-parametric.}
\description{
An S4 class to represent a distribution, either parametric or non-parametric.
}
\section{Slots}{

\describe{
\item{\code{p1}}{A number for the value of the first parameter (default: 0.8).}

\item{\code{p2}}{A number for the value of the second parameter (default: 1).}

\item{\code{p3}}{A number for the value of the third parameter (default: 0).}

\item{\code{empirical}}{A matrix that defines an empirical distribution with values and probabilities.}

\item{\code{min}}{A number that defines the minimum value of the variable (default: 1e-8 considering it is used for frequency and severity modeling).}

\item{\code{max}}{A number that defines the maximum value of the variable (default: 1e8).}

\item{\code{fitsucc}}{Whether a distribution fitting is successful.}

\item{\code{info}}{A character string that contains additional information of the distribution to identify line/type/frequency or severity.}
}}

