% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R, R/Distribution.R
\docType{methods}
\name{doPlot}
\alias{doPlot}
\alias{doPlot,Distribution-method}
\alias{doPlot,ANY-method}
\title{Plot function.}
\usage{
doPlot(object, ...)

\S4method{doPlot}{Distribution}(object)
}
\arguments{
\item{object}{Object}

\item{...}{Additional function arguments}
}
\description{
Plot function.
}
\examples{
xPareto <- new("Pareto",p1=20,p2=3)
doPlot(xPareto)
}
