% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{doPlot}
\alias{doPlot}
\title{Plot function.}
\usage{
doPlot(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional function arguments}
}
\description{
Plot function.
}
\examples{
xPareto <- new("Pareto",p1=20,p2=3)
doPlot(xPareto)
}
