% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{doSample}
\alias{doSample}
\title{Sampling from the distribution.}
\usage{
doSample(object, n, ...)
}
\arguments{
\item{object}{A Distribution Object}

\item{n}{Number of samples}
}
\description{
Sampling from the distribution.
}
\examples{
xPareto <- new("Pareto",p1=20,p2=3)
doSample(xPareto,10000)
}
