% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{expectZeros}
\alias{expectZeros}
\title{Get the expected P0 based on settlement/close year.}
\usage{
expectZeros(closeYear, zeroProb)
}
\arguments{
\item{closeYear}{Development years that claims are settled. It could be a number or a numeric vector.}

\item{zeroProb}{A vector that contains the P(0) based on development year.}
}
\description{
Get the expected P0 based on settlement/close year.
}
\examples{
zeroprob<-c(0.02,0.01,0.005,0.005,0.003,0)
expectZeros(c(2,3,6,9,100,1,2,3,4),zeroprob)
}
