% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{fitPlot}
\alias{fitPlot}
\alias{fitPlot,FitDist-method}
\alias{fitPlot,ANY-method}
\title{Compare the raw data and fitted distribution on density, CDF, Q-Q plot and P-P plot}
\usage{
fitPlot(object, ...)

\S4method{fitPlot}{FitDist}(object, n = missing)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}

\item{n}{Number of samples, should not be used in current setting}
}
\description{
Compare the raw data and fitted distribution on density, CDF, Q-Q plot and P-P plot
}
