% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{fitPlot}
\alias{fitPlot}
\title{Compare the raw data and fitted distribution on density, CDF, Q-Q plot and P-P plot}
\usage{
fitPlot(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Compare the raw data and fitted distribution on density, CDF, Q-Q plot and P-P plot
}
