% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{fitStartValue}
\alias{fitStartValue}
\title{Set the initial parameter values of distribution fitting.}
\usage{
fitStartValue(object, ...)
}
\arguments{
\item{object}{Distribution Object}
}
\description{
Set the initial parameter values of distribution fitting.
}
