% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{getAIC}
\alias{getAIC}
\title{Get the AIC of distribution fitting.}
\usage{
getAIC(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Get the AIC of distribution fitting.
}
