% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{getBIC}
\alias{getBIC}
\title{Get the BIC of distribution fitting.}
\usage{
getBIC(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Get the BIC of distribution fitting.
}
