% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula.R
\docType{methods}
\name{getCopula}
\alias{getCopula}
\alias{getCopula,CopulaObj-method}
\alias{getCopula,ANY-method}
\title{Get the R copula object.}
\usage{
getCopula(object, ...)

\S4method{getCopula}{CopulaObj}(object)
}
\arguments{
\item{object}{R copula object}

\item{...}{Additional parameters that may or may not be used}
}
\description{
Get the R copula object.
}
