% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{getFitdata}
\alias{getFitdata}
\title{getFitdata
Get processed data used for distribution fitting.}
\usage{
getFitdata(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
getFitdata
Get processed data used for distribution fitting.
}
