% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{getfitmethod}
\alias{getfitmethod}
\title{Get the distribution fitting method: mle, mme or qme.}
\usage{
getfitmethod(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Get the distribution fitting method: mle, mme or qme.
}
