% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{getFittedDist}
\alias{getFittedDist}
\title{Get fitted distribution.}
\usage{
getFittedDist(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Get fitted distribution.
}
