% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{getfoutput}
\alias{getfoutput}
\title{Get frequency distribution fitting summary.}
\usage{
getfoutput(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Get frequency distribution fitting summary.
}
