% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{getFreq}
\alias{getFreq}
\title{Get the frequency of data: Annual or Monthly. Only used for frequency fitting.}
\usage{
getFreq(object, ...)
}
\arguments{
\item{object}{FitDist Object}
}
\description{
Get the frequency of data: Annual or Monthly. Only used for frequency fitting.
}
