% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{getiDate}
\alias{getiDate}
\title{Check if occurrence dates are used (TRUE) or not (FALSE) for frequency fitting.}
\usage{
getiDate(object, ...)
}
\arguments{
\item{object}{FitDist Object}
}
\description{
Check if occurrence dates are used (TRUE) or not (FALSE) for frequency fitting.
}
