% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIndex.R
\name{getIndex}
\alias{getIndex}
\title{Retrieve index value based on dates.}
\usage{
getIndex(object, ...)
}
\arguments{
\item{object}{Index Object}
}
\description{
\code{getIndex} sets a time index to reflect inflation, underwriting cycle or seasonality.
}
\examples{
xindex <- new("Index", indexID = "IDX1", tabulate = FALSE, annualizedRate = 0.03)
xindex<-setIndex(xindex)
xindex@monthlyIndex
dates<-as.Date("2015-12-31")
getIndex(xindex,dates)
}
