% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{getMean}
\alias{getMean}
\title{Get the mean of experience data used for fitting.}
\usage{
getMean(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Get the mean of experience data used for fitting.
}
