% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R, R/FitDist.R
\docType{methods}
\name{getObservation}
\alias{getObservation}
\alias{getObservation,FitDist-method}
\alias{getObservation,ANY-method}
\title{Get input data from an object.}
\usage{
getObservation(object, ...)

\S4method{getObservation}{FitDist}(object)
}
\arguments{
\item{object}{Object}

\item{...}{Additional function arguments}
}
\description{
Get input data from an object.
}
