% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{getObservation}
\alias{getObservation}
\title{Get input data from an object.}
\usage{
getObservation(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional function arguments}
}
\description{
Get input data from an object.
}
