% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{getPSD}
\alias{getPSD}
\title{Get the standard deviation of parameter estimation in distribution fitting. It is only available for mle.}
\usage{
getPSD(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Get the standard deviation of parameter estimation in distribution fitting. It is only available for mle.
}
