% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{getSd}
\alias{getSd}
\title{Get the standard deviation of experience data used for fitting.}
\usage{
getSd(object, ...)
}
\arguments{
\item{object}{FitDist Object}
}
\description{
Get the standard deviation of experience data used for fitting.
}
