% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{getsoutput}
\alias{getsoutput}
\title{Get severity distribution fitting summary.}
\usage{
getsoutput(object, ...)
}
\arguments{
\item{object}{FitDist Object}
}
\description{
Get severity distribution fitting summary.
}
