% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R, R/FitDist.R
\docType{methods}
\name{getTrend}
\alias{getTrend}
\alias{getTrend,FitDist-method}
\alias{getTrend,ANY-method}
\title{Get the trend index.}
\usage{
getTrend(object, ...)

\S4method{getTrend}{FitDist}(object)
}
\arguments{
\item{object}{Object}

\item{...}{Additional function arguments}
}
\description{
Get the trend index.
}
