% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{nloglik}
\alias{nloglik}
\title{Negative Loglikelihood.}
\usage{
nloglik(paras, dist, fitdata, deductible, limit)
}
\arguments{
\item{paras}{A vector contain distribution parameters.}

\item{dist}{A Distribution Object.}

\item{fitdata}{A vector of loss data for fitting.}

\item{deductible}{A vector of deductible data for all loss data.}

\item{limit}{A vector of limit data for all loss data.}
}
\description{
Negative Loglikelihood.
}
\examples{
paras<-c(1,1)
dist<-new("Normal")
fitdata<-rtnorm(1000,3,2,1,10)
deductible<-rep(1,1000)
limit<-rep(9,1000)
nloglik(paras,dist,fitdata,deductible,limit)
paras<-c(3,2)
nloglik(paras,dist,fitdata,deductible,limit)
}
