% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{nParameter}
\alias{nParameter}
\title{Set the number of parameters for distribution fitting.}
\usage{
nParameter(object, ...)
}
\arguments{
\item{object}{Distribution Object}
}
\description{
Set the number of parameters for distribution fitting.
}
