% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{observationPlot}
\alias{observationPlot}
\alias{observationPlot,FitDist-method}
\alias{observationPlot,ANY-method}
\title{Plotting the data for distribution fitting}
\usage{
observationPlot(object, ...)

\S4method{observationPlot}{FitDist}(object)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Plotting the data for distribution fitting
}
