% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{observationPlot}
\alias{observationPlot}
\title{Plotting the data for distribution fitting}
\usage{
observationPlot(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Plotting the data for distribution fitting
}
