% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{PDFPlot}
\alias{PDFPlot}
\alias{PDFPlot,FitDist-method}
\alias{PDFPlot,ANY-method}
\title{Plotting the PDF of data and fitted distribution}
\usage{
PDFPlot(object, ...)

\S4method{PDFPlot}{FitDist}(object, n = missing)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}

\item{n}{Number of samples, should not be used in current setting}
}
\description{
Plotting the PDF of data and fitted distribution
}
