% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{PPPlot}
\alias{PPPlot}
\alias{PPPlot,FitDist-method}
\alias{PPPlot,ANY-method}
\title{P-P Plot of data and fitted distribution}
\usage{
PPPlot(object, ...)

\S4method{PPPlot}{FitDist}(object, n = missing)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}

\item{n}{Number of samples, should not be used in current setting}
}
\description{
P-P Plot of data and fitted distribution
}
