% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{Probability}
\alias{Probability}
\title{Probability function.}
\usage{
Probability(object, q, ...)
}
\arguments{
\item{object}{Distribution Object}

\item{q}{Variable value}

\item{...}{Additional function arguments}
}
\description{
Probability function.
}
\examples{
xPareto <- new("Pareto",p1=20,p2=3)
Probability(xPareto,50)
}
