% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{QQPlot}
\alias{QQPlot}
\alias{QQPlot,FitDist-method}
\alias{QQPlot,ANY-method}
\title{Q-Q Plot of data and fitted distribution}
\usage{
QQPlot(object, ...)

\S4method{QQPlot}{FitDist}(object, n = missing)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}

\item{n}{Number of samples, should not be used in current setting}
}
\description{
Q-Q Plot of data and fitted distribution
}
