% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{Quantile}
\alias{Quantile}
\title{Quantile function.}
\usage{
Quantile(object, p, ...)
}
\arguments{
\item{object}{Distribution Object}

\item{p}{Probability}

\item{...}{Additional function arguments}
}
\description{
Quantile function.
}
\examples{
xPareto <- new("Pareto",p1=20,p2=3)
Quantile(xPareto,0.6)
}
