% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{rreopen}
\alias{rreopen}
\title{Simulate whether closed claims will be reopened or not.}
\usage{
rreopen(closeYear, reopenProb)
}
\arguments{
\item{closeYear}{Years after claim closure. It could be a number or a numeric vector.}

\item{reopenProb}{A vector that contains the reopen probability based on closeYear.}
}
\description{
Simulate whether closed claims will be reopened or not.
}
\examples{
reopenprob<-c(0.02,0.01,0.005,0.005,0.003,0)
rreopen(rep(2,1000),reopenprob)
}
