% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{methods}
\name{sampleKurtosis}
\alias{sampleKurtosis}
\alias{sampleKurtosis,Distribution-method}
\alias{sampleKurtosis,ANY-method}
\title{Calculate the excess kurtosis of 10000 sampled values from the distribution.}
\usage{
sampleKurtosis(object, ...)

\S4method{sampleKurtosis}{Distribution}(object)
}
\arguments{
\item{object}{A Distribution Object}

\item{...}{Additional function arguments}
}
\description{
Calculate the excess kurtosis of 10000 sampled values from the distribution.
}
\examples{
xLognormal <- new("Lognormal",p1=2,p2=3)
sampleKurtosis(xLognormal)
}
