% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{sampleKurtosis}
\alias{sampleKurtosis}
\title{Calculate the excess kurtosis of 10000 sampled values from the distribution.}
\usage{
sampleKurtosis(object, ...)
}
\arguments{
\item{object}{A Distribution Object}
}
\description{
Calculate the excess kurtosis of 10000 sampled values from the distribution.
}
\examples{
xLognormal <- new("Lognormal",p1=2,p2=3)
sampleKurtosis(xLognormal)
}
