% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{methods}
\name{sampleMean}
\alias{sampleMean}
\alias{sampleMean,Distribution-method}
\alias{sampleMean,ANY-method}
\title{Calculate the mean of 100000 sampled values from the distribution.}
\usage{
sampleMean(object, ...)

\S4method{sampleMean}{Distribution}(object)
}
\arguments{
\item{object}{A Distribution Object}

\item{...}{Additional function arguments}
}
\description{
Calculate the mean of 100000 sampled values from the distribution.
}
\examples{
xLognormal <- new("Lognormal",p1=2,p2=3)
sampleMean(xLognormal)
}
