% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{methods}
\name{sampleSd}
\alias{sampleSd}
\alias{sampleSd,Distribution-method}
\alias{sampleSd,ANY-method}
\title{Calculate the standard deviation of 10000 sampled values from the distribution.}
\usage{
sampleSd(object, ...)

\S4method{sampleSd}{Distribution}(object)
}
\arguments{
\item{object}{A Distribution Object}

\item{...}{Additional function arguments}
}
\description{
Calculate the standard deviation of 10000 sampled values from the distribution.
}
\examples{
xLognormal <- new("Lognormal",p1=2,p2=3)
sampleSd(xLognormal)
}
