% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{sampleSkew}
\alias{sampleSkew}
\title{Calculate the skewness of 10000 sampled values from the distribution.}
\usage{
sampleSkew(object, ...)
}
\arguments{
\item{object}{A Distribution Object}
}
\description{
Calculate the skewness of 10000 sampled values from the distribution.
}
\examples{
xLognormal <- new("Lognormal",p1=2,p2=3)
sampleSkew(xLognormal)
}
