% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIndex.R
\docType{methods}
\name{setAnnualizedRate<-}
\alias{setAnnualizedRate<-}
\alias{setAnnualizedRate<-,Index,numeric-method}
\alias{setAnnualizedRate,ANY-method}
\title{Set the annualized level rate to construct the index.
Only used when tabulate == FALSE.}
\usage{
setAnnualizedRate(this, ...) <- value

\S4method{setAnnualizedRate}{Index,numeric}(this) <- value
}
\arguments{
\item{this}{Index Object}

\item{...}{Additional function arguments}

\item{value}{Numeric Value (default:0.02)}
}
\description{
Set the annualized level rate to construct the index.
Only used when tabulate == FALSE.
}
\examples{
xindex <- new("Index")
setID(xindex)<-"IDX1"
setTabulate(xindex)<-FALSE
setAnnualizedRate(xindex)<-0.03
xindex<-setIndex(xindex)
xindex@monthlyIndex
}
