% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIndex.R
\name{setAnnualizedRate<-}
\alias{setAnnualizedRate<-}
\title{Set the annualized level rate to construct the index.
Only used when tabulate == FALSE.}
\usage{
setAnnualizedRate(this, ...) <- value
}
\arguments{
\item{this}{Index Object}

\item{value}{Numeric Value (default:0.02)}
}
\description{
Set the annualized level rate to construct the index.
Only used when tabulate == FALSE.
}
