% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula.R
\docType{methods}
\name{setCopulaParam<-}
\alias{setCopulaParam<-}
\alias{setCopulaParam<-,CopulaObj,numeric-method}
\alias{setCopulaParam,ANY-method}
\title{Set copula parameters.}
\usage{
setCopulaParam(this, ...) <- value

\S4method{setCopulaParam}{CopulaObj,numeric}(this) <- value
}
\arguments{
\item{this}{Copula Object}

\item{...}{Additional function arguments}

\item{value}{The copula parameters}
}
\description{
Set copula parameters.
}
