% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula.R
\docType{methods}
\name{setCopulaType<-}
\alias{setCopulaType<-}
\alias{setCopulaType<-,CopulaObj,character-method}
\alias{setCopulaType,ANY-method}
\title{Set copula type.}
\usage{
setCopulaType(this, ...) <- value

\S4method{setCopulaType}{CopulaObj,character}(this) <- value
}
\arguments{
\item{this}{Copula Object}

\item{...}{Additional function arguments}

\item{value}{The copula type}
}
\description{
Set copula type.
}
