% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula.R
\docType{methods}
\name{setDimension<-}
\alias{setDimension<-}
\alias{setDimension<-,CopulaObj,numeric-method}
\alias{setDimension,ANY-method}
\title{Set the dimension of the copula.}
\usage{
setDimension(this, ...) <- value

\S4method{setDimension}{CopulaObj,numeric}(this) <- value
}
\arguments{
\item{this}{Copula Object}

\item{...}{Additional function arguments}

\item{value}{The dimension of the copula. It can also be set by providing marginal distributions}
}
\description{
Set the dimension of the copula.
}
