% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{methods}
\name{setEmpirical<-}
\alias{setEmpirical<-}
\alias{setEmpirical<-,Distribution,matrix-method}
\alias{setEmpirical,ANY-method}
\title{Set the list of values and corresponding probabilities (Pr(X<value) for continuous variable and Pr(X==value) for discrete variable).
It is only used for empirical distribution.}
\usage{
setEmpirical(this, ...) <- value

\S4method{setEmpirical}{Distribution,matrix}(this) <- value
}
\arguments{
\item{this}{Distribution Object}

\item{...}{Additional function arguments.}

\item{value}{Two-column matrix with values and probabilities}
}
\description{
Set the list of values and corresponding probabilities (Pr(X<value) for continuous variable and Pr(X==value) for discrete variable).
It is only used for empirical distribution.
}
