% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFac.R
\docType{methods}
\name{setFacModel<-}
\alias{setFacModel<-}
\alias{setFacModel<-,DevFac,logical-method}
\alias{setFacModel,ANY-method}
\title{Determine whether the development factor is determined by a predictive model or a fixed schedule by development year}
\usage{
setFacModel(this, ...) <- value

\S4method{setFacModel}{DevFac,logical}(this) <- value
}
\arguments{
\item{this}{DevFac Object}

\item{...}{Additional function arguments}

\item{value}{Logical Value (default:FALSE)}
}
\description{
Determine whether the development factor is determined by a predictive model or a fixed schedule by development year
}
