% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{setFitdata}
\alias{setFitdata}
\alias{setFitdata,FitDist-method}
\alias{setFitdata,ANY-method}
\title{Preparing the input data (observation) for distribution fitting, including detrending, translating occurrence dates to frequency data, etc.}
\usage{
setFitdata(object, ...)

\S4method{setFitdata}{FitDist}(object)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Preparing the input data (observation) for distribution fitting, including detrending, translating occurrence dates to frequency data, etc.
}
