% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{setFitdata}
\alias{setFitdata}
\title{Preparing the input data (observation) for distribution fitting, including detrending, translating occurrence dates to frequency data, etc.}
\usage{
setFitdata(object, ...)
}
\arguments{
\item{object}{FitDist Object}
}
\description{
Preparing the input data (observation) for distribution fitting, including detrending, translating occurrence dates to frequency data, etc.
}
