% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{setfitmethod<-}
\alias{setfitmethod<-}
\alias{setfitmethod<-,FitDist,character-method}
\alias{setfitmethod,ANY-method}
\title{Set distribution fitting method.}
\usage{
setfitmethod(this, ...) <- value

\S4method{setfitmethod}{FitDist,character}(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{...}{Additional function arguments}

\item{value}{A character string: "mle", "mme", or "qme"}
}
\description{
Set distribution fitting method.
}
