% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{setfitmethod<-}
\alias{setfitmethod<-}
\title{Set distribution fitting method.}
\usage{
setfitmethod(this, ...) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{A character string: "mle", "mme", or "qme"}
}
\description{
Set distribution fitting method.
}
