% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{setFittedDist<-}
\alias{setFittedDist<-}
\alias{setFittedDist<-,FitDist,Distribution-method}
\alias{setFittedDist,ANY-method}
\title{Directly set the fitted distribution without fitting it to the data.}
\usage{
setFittedDist(this) <- value

\S4method{setFittedDist}{FitDist,Distribution}(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{Fitted distribution}
}
\description{
Directly set the fitted distribution without fitting it to the data.
}
