% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{setFittedDist<-}
\alias{setFittedDist<-}
\title{Directly set the fitted distribution without fitting it to the data.}
\usage{
setFittedDist(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{Fitted distribution}
}
\description{
Directly set the fitted distribution without fitting it to the data.
}
