% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{setfreq<-}
\alias{setfreq<-}
\alias{setfreq<-,FitDist,character-method}
\alias{setfreq,ANY-method}
\title{Set the data frequency.}
\usage{
setfreq(this, ...) <- value

\S4method{setfreq}{FitDist,character}(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{...}{Additional function arguments}

\item{value}{A character string: "Annual" or "Monthly"}
}
\description{
Set the data frequency.
}
