% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{setfreq<-}
\alias{setfreq<-}
\title{Set the data frequency.}
\usage{
setfreq(this, ...) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{A character string: "Annual" or "Monthly"}
}
\description{
Set the data frequency.
}
