% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFac.R
\docType{methods}
\name{setFun<-}
\alias{setFun<-}
\alias{setFun<-,DevFac,character-method}
\alias{setFun,ANY-method}
\title{Set the model format/link function (identity/inverse/log/exponential).
Only used when FacModel == TRUE.}
\usage{
setFun(this, ...) <- value

\S4method{setFun}{DevFac,character}(this) <- value
}
\arguments{
\item{this}{DevFac Object}

\item{...}{Additional function arguments}

\item{value}{String Value (default:"identity")}
}
\description{
Set the model format/link function (identity/inverse/log/exponential).
Only used when FacModel == TRUE.
}
