% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFac.R
\name{setFun<-}
\alias{setFun<-}
\title{Set the model format/link function (identity/inverse/log/exponential).
Only used when FacModel == TRUE.}
\usage{
setFun(this, ...) <- value
}
\arguments{
\item{this}{DevFac Object}

\item{value}{String Value (default:"identity")}
}
\description{
Set the model format/link function (identity/inverse/log/exponential).
Only used when FacModel == TRUE.
}
