% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R, R/CIndex.R, R/DevFac.R
\docType{methods}
\name{setID<-}
\alias{setID<-}
\alias{setID<-,Index,character-method}
\alias{setID,ANY-method}
\alias{setID<-,DevFac,character-method}
\title{setID
Set the ID for an object}
\usage{
setID(this, ...) <- value

\S4method{setID}{Index,character}(this) <- value

\S4method{setID}{DevFac,character}(this) <- value
}
\arguments{
\item{this}{Self}

\item{...}{Additional function arguments}

\item{value}{ID}
}
\description{
setID
Set the ID for an object
}
\examples{
xindex <- new("Index")
setID(xindex)<-"IDX1"
xindex@indexID
}
