% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{setID<-}
\alias{setID<-}
\title{setID
Set the ID for an object}
\usage{
setID(this, ...) <- value
}
\arguments{
\item{this}{Self}

\item{...}{Additional function arguments}

\item{value}{ID}
}
\description{
setID
Set the ID for an object
}
