% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{setidate<-}
\alias{setidate<-}
\alias{setidate<-,FitDist,logical-method}
\alias{setidate,ANY-method}
\title{Set whether occurrence dates will be used for frequency data.}
\usage{
setidate(this, ...) <- value

\S4method{setidate}{FitDist,logical}(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{...}{Additional function arguments}

\item{value}{A boolean}
}
\description{
Set whether occurrence dates will be used for frequency data.
}
