% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{setidate<-}
\alias{setidate<-}
\title{Set whether occurrence dates will be used for frequency data.}
\usage{
setidate(this, ...) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{A boolean}
}
\description{
Set whether occurrence dates will be used for frequency data.
}
