% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{setifreq<-}
\alias{setifreq<-}
\alias{setifreq<-,FitDist,logical-method}
\alias{setifreq,ANY-method}
\title{Set the data type: frequency or severity/time lag.}
\usage{
setifreq(this, ...) <- value

\S4method{setifreq}{FitDist,logical}(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{...}{Additional function arguments}

\item{value}{A boolean}
}
\description{
Set the data type: frequency or severity/time lag.
}
