% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula.R
\docType{methods}
\name{setMarginal<-}
\alias{setMarginal<-}
\alias{setMarginal<-,CopulaObj,list-method}
\alias{setMarginal,ANY-method}
\title{Set the marginal distributions of the copula.}
\usage{
setMarginal(this, ...) <- value

\S4method{setMarginal}{CopulaObj,list}(this) <- value
}
\arguments{
\item{this}{Copula Object}

\item{...}{Additional function arguments}

\item{value}{The list of marginal distributions.}
}
\description{
Set the marginal distributions of the copula.
}
