% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFac.R
\name{setMeanList<-}
\alias{setMeanList<-}
\title{Set the year-to-year loss development factor.}
\usage{
setMeanList(this, ...) <- value
}
\arguments{
\item{this}{DevFac Object}

\item{value}{Numeric Vector}
}
\description{
\code{setMeanList<-} sets expected year-to-year loss development factor. Years after It is only used when ibnerfModel == FALSE.
}
