% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{setMin}
\alias{setMin}
\title{Set the minimum of the distribution. For example, the distribution of settlement lag for open claims}
\usage{
setMin(object, ...)
}
\arguments{
\item{object}{A Distribution Object}

\item{minval}{The minimum value.}
}
\description{
Set the minimum of the distribution. For example, the distribution of settlement lag for open claims
}
\examples{
xLognormal <- new("Lognormal",p1=2,p2=3)
xLognormal <- setMin(xLognormal,50)
}
