% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{setObservation<-}
\alias{setObservation<-}
\title{Input the raw data.}
\usage{
setObservation(this) <- value
}
\arguments{
\item{this}{FitDist Object or Copula Object}

\item{value}{A data frame or a matrix. For FitDist object, it could be a two-column data frame with the occurrence date and loss size/number of occurrence. Or a one-column data frame with loss size (ifreq == FALSE) or number of occurrence (ifreq == TRUE && idate == FALSE) or occurrence dates (ifreq == TRUE && idate == TRUE). For Copula object, it could be a matrix with each column contains the experience data of a variable.}
}
\description{
Input the raw data.
}
