% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{methods}
\name{setParams<-}
\alias{setParams<-}
\alias{setParams<-,Distribution,numeric-method}
\alias{setParams,ANY-method}
\title{Set distribution parameters.}
\usage{
setParams(this, ...) <- value

\S4method{setParams}{Distribution,numeric}(this) <- value
}
\arguments{
\item{this}{Distribution Object}

\item{...}{Additional function arguments.}

\item{value}{Numeric vector containing parameters
examples
dist <- new("Normal")
setParams(dist) <- c(2,3)
dist}
}
\description{
Set distribution parameters.
}
