% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFac.R
\name{setParas<-}
\alias{setParas<-}
\title{Set the values of model parameters.}
\usage{
setParas(this, ...) <- value
}
\arguments{
\item{this}{DevFac Object}

\item{value}{Numeric Vector}
}
\description{
\code{setParas<-} sets model parameters. Their order must match the order of c("Intercept","DevelopmentYear","IncurredLoss","OSRatio",xname,"Volatility"). "Volatility" stands for the volatility of the error term in the model and used to simulate IBNER development factors. The parameter vector is only used when ibnerfModel == TRUE.
}
