% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{setprobs<-}
\alias{setprobs<-}
\title{Set the percentiles to be matched. Only used when qme is chosen for fitting method.}
\usage{
setprobs(this, ...) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{A numeric vector with values between 0 and 1.}
}
\description{
Set the percentiles to be matched. Only used when qme is chosen for fitting method.
}
