% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{methods}
\name{setRange<-}
\alias{setRange<-}
\alias{setRange<-,Distribution,numeric-method}
\alias{setRange,ANY-method}
\title{Set the min and max of the variable.}
\usage{
setRange(this, ...) <- value

\S4method{setRange}{Distribution,numeric}(this) <- value
}
\arguments{
\item{this}{Distribution Object}

\item{...}{Additional function arguments.}

\item{value}{a two-element vector contains min and max.}
}
\description{
Set the min and max of the variable.
}
