% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{setRange<-}
\alias{setRange<-}
\title{Set the min and max of the variable.}
\usage{
setRange(this, ...) <- value
}
\arguments{
\item{this}{Distribution Object}

\item{value}{a two-element vector contains min and max.}
}
\description{
Set the min and max of the variable.
}
