% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTriangle.R
\docType{methods}
\name{setRectangle}
\alias{setRectangle}
\alias{setRectangle,Triangle,data.frame-method}
\alias{setRectangle,ANY-method}
\title{Set up the rectangle based on simulated data.}
\usage{
setRectangle(object, data, ...)

\S4method{setRectangle}{Triangle,data.frame}(object, data,
  evaluationDate = as.Date("2016-12-31"),
  futureDate = as.Date("2017-12-31"), lob = "Total", ctype = "Total")
}
\arguments{
\item{object}{Triangle Object}

\item{data}{Simulated Data}

\item{...}{Additional function arguments.}

\item{evaluationDate}{Evaluation Date;}

\item{futureDate}{End of projection date;}

\item{lob}{Line of Business;}

\item{ctype}{Claim Type.}
}
\description{
\code{setRectangle} sets up the rectangle based on a data file.
}
