% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTriangle.R
\name{setRectangle}
\alias{setRectangle}
\title{Set up the rectangle based on simulated data.}
\usage{
setRectangle(object, data, ...)
}
\arguments{
\item{object}{Triangle Object}

\item{data}{Simulated Data}

\item{evaluationDate}{Evaluation Date}

\item{lob}{Line of Business}

\item{ctype}{Claim Type}
}
\description{
\code{setRectangle} sets up the rectangle based on a data file.
}
