% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIndex.R
\docType{methods}
\name{setSeasonality<-}
\alias{setSeasonality<-}
\alias{setSeasonality<-,Index,vector-method}
\alias{setSeasonality,ANY-method}
\title{Set seasonality on a monthly basis.}
\usage{
setSeasonality(this, ...) <- value

\S4method{setSeasonality}{Index,vector}(this) <- value
}
\arguments{
\item{this}{Index Object}

\item{...}{Additional function arguments}

\item{value}{Numeric Vector (default:rep(1,12))}
}
\description{
\code{setSeasonality<-} sets monthly multiplier to reflect seasonal impact.
}
