% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIndex.R
\name{setSeasonality<-}
\alias{setSeasonality<-}
\title{Set seasonality on a monthly basis.}
\usage{
setSeasonality(this, ...) <- value
}
\arguments{
\item{this}{Index Object}

\item{value}{Numeric Vector (default:rep(1,12))}
}
\description{
\code{setSeasonality<-} sets monthly multiplier to reflect seasonal impact.
}
