% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R, R/CIndex.R
\docType{methods}
\name{setStartDate<-}
\alias{setStartDate<-}
\alias{setStartDate<-,Index,Date-method}
\alias{setStartDate,ANY-method}
\title{Set the start date for the claim simulation exercise}
\usage{
setStartDate(this, ...) <- value

\S4method{setStartDate}{Index,Date}(this) <- value
}
\arguments{
\item{this}{Self}

\item{...}{Additional function arguments}

\item{value}{Start Date}
}
\description{
Set the start date for the claim simulation exercise
}
