% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{setStartDate<-}
\alias{setStartDate<-}
\title{Set the start date for the claim simulation exercise}
\usage{
setStartDate(this, ...) <- value
}
\arguments{
\item{this}{Self}

\item{...}{Additional function arguments}

\item{value}{Start Date}
}
\description{
Set the start date for the claim simulation exercise
}
