% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIndex.R
\name{setTabulate<-}
\alias{setTabulate<-}
\title{Determine whether the index values are constructed from a constant rate or provided directly}
\usage{
setTabulate(this, ...) <- value
}
\arguments{
\item{this}{Index Object}

\item{value}{Logical Value (default:FALSE)}
}
\description{
Determine whether the index values are constructed from a constant rate or provided directly
}
