% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{setTrend<-}
\alias{setTrend<-}
\alias{setTrend<-,FitDist,Index-method}
\alias{setTrend,ANY-method}
\title{Set the trend with an Index Object.}
\usage{
setTrend(this, ...) <- value

\S4method{setTrend}{FitDist,Index}(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{...}{Additional function arguments}

\item{value}{An Index Object}
}
\description{
Set the trend with an Index Object.
}
