% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{setTrend<-}
\alias{setTrend<-}
\title{Set the trend with an Index Object.}
\usage{
setTrend(this, ...) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{An Index Object}
}
\description{
Set the trend with an Index Object.
}
