% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{setTrialDist<-}
\alias{setTrialDist<-}
\alias{setTrialDist<-,FitDist,Distribution-method}
\alias{setTrialDist,ANY-method}
\title{Distribution fitting and testing.}
\usage{
setTrialDist(this) <- value

\S4method{setTrialDist}{FitDist,Distribution}(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{Distribution to fit to}
}
\description{
Distribution fitting and testing.
}
