% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{setTrialDist<-}
\alias{setTrialDist<-}
\title{Distribution fitting and testing.}
\usage{
setTrialDist(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{Distribution to fit to}
}
\description{
Distribution fitting and testing.
}
