% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\docType{methods}
\name{setTrialDistErr<-}
\alias{setTrialDistErr<-}
\alias{setTrialDistErr<-,FitDist,Distribution-method}
\alias{setTrialDistErr,ANY-method}
\title{Distribution fitting and testing. Same as setTrialDist except for error tolerance.}
\usage{
setTrialDistErr(this) <- value

\S4method{setTrialDistErr}{FitDist,Distribution}(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{Distribution to fit to}
}
\description{
Distribution fitting and testing. Same as setTrialDist except for error tolerance.
}
