% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{setTrialDistErr<-}
\alias{setTrialDistErr<-}
\title{Distribution fitting and testing. Same as setTrialDist except for error tolerance.}
\usage{
setTrialDistErr(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{Distribution to fit to}
}
\description{
Distribution fitting and testing. Same as setTrialDist except for error tolerance.
}
