% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{setTruncated<-}
\alias{setTruncated<-}
\title{Set the indicator of truncated distribution.}
\usage{
setTruncated(this, ...) <- value
}
\arguments{
\item{this}{Distribution Object}

\item{value}{Boolean to indicate whether the distribution is truncated by min and max or not.}
}
\description{
Set the indicator of truncated distribution.
}
