% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTriangle.R
\name{setUpperTriangle}
\alias{setUpperTriangle}
\title{Set up the upper triangle based on claim data.}
\usage{
setUpperTriangle(object, data, ...)
}
\arguments{
\item{object}{Triangle Object}

\item{data}{Claim Data}

\item{evaluationDate}{Evaluation Date}

\item{lob}{Line of Business}

\item{ctype}{Claim Type}
}
\description{
\code{setUpperTriangle} sets up the upper triangle based on a data file.
}
\examples{
library(cascsim)
data(claimdata)
xTri <- new("Triangle", triID = "TRI1", type = "reportedCount", startDate=as.Date("2012-01-01"),
frequency="yearly", sim=1, percentile=50)
xTri<-setUpperTriangle(xTri,claimdata)
xTri@upper

xTri <- new("Triangle", triID = "TRI1", type = "closedCount", startDate=as.Date("2012-01-01"),
frequency="quarterly", sim=1, percentile=50)
xTri<-setUpperTriangle(xTri,claimdata)
xTri@upper

xTri <- new("Triangle", triID = "TRI1", type = "incurredLoss", startDate=as.Date("2012-01-01"),
frequency="yearly", sim=1, percentile=50)
xTri<-setUpperTriangle(xTri,claimdata,lob="Auto",ctype="H")
xTri@upper

xTri <- new("Triangle", triID = "TRI1", type = "paidLoss", startDate=as.Date("2012-01-01"),
frequency="yearly", sim=1, percentile=50)
xTri<-setUpperTriangle(xTri,claimdata,lob="Auto",ctype="H")
xTri@upper

}
