% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFac.R
\docType{methods}
\name{setVolList<-}
\alias{setVolList<-}
\alias{setVolList<-,DevFac,vector-method}
\alias{setVolList,ANY-method}
\title{Set the year-to-year loss development factor volatility.}
\usage{
setVolList(this, ...) <- value

\S4method{setVolList}{DevFac,vector}(this) <- value
}
\arguments{
\item{this}{DevFac Object}

\item{...}{Additional function arguments}

\item{value}{Numeric Vector}
}
\description{
\code{setMeanList<-} sets year-to-year loss development factor volatility. It is used to simulate loss development factor assuming a normal distribution. It can be set to zero for deterministic estimation. It is only used when ibnerfModel == FALSE.
}
