% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFac.R
\name{setVolList<-}
\alias{setVolList<-}
\title{Set the year-to-year loss development factor volatility.}
\usage{
setVolList(this, ...) <- value
}
\arguments{
\item{this}{DevFac Object}

\item{value}{Numeric Vector}
}
\description{
\code{setMeanList<-} sets year-to-year loss development factor volatility. It is used to simulate loss development factor assuming a normal distribution. It can be set to zero for deterministic estimation. It is only used when ibnerfModel == FALSE.
}
