% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIndex.R
\name{shiftIndex}
\alias{shiftIndex}
\title{Shift monthly index with a new start date and replace the unknown index value with zero.}
\usage{
shiftIndex(object, ...)
}
\arguments{
\item{object}{Index Object}
}
\description{
Shift monthly index with a new start date and replace the unknown index value with zero.
}
\examples{
xindex <- new("Index", indexID = "IDX1", tabulate = FALSE, annualizedRate = 0.03)
xindex<-setIndex(xindex)
xindex@monthlyIndex
shiftIndex(xindex,as.Date("2016-10-15"))
shiftIndex(xindex,as.Date("2010-10-15"))
}
