% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\docType{methods}
\name{simReport}
\alias{simReport}
\alias{simReport,Simulation,data.frame-method}
\alias{simReport,ANY-method}
\title{Generate claim simulation result report in html}
\usage{
simReport(object, simSummary, ...)

\S4method{simReport}{Simulation,data.frame}(object, simSummary,
  simTriangle = NA, startDate = as.Date("2012-01-01"),
  evaluationDate = as.Date("2016-12-31"),
  futureDate = as.Date("2017-12-31"), iYear = FALSE)
}
\arguments{
\item{object}{ClaimType object}

\item{simSummary}{simulation result summary generated by simSummary}

\item{...}{Additional parameters that may or may not be used.}

\item{simTriangle}{triangle summary generated by simTriangle;}

\item{startDate}{Date after which claims are analyzed;}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR;}

\item{futureDate}{Date of evaluation for UPR (future claims);}

\item{iYear}{Boolean that indicates whether summary by accident year should be produced in the report.}
}
\description{
Generate claim simulation result report in html
}
