% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{simReport}
\alias{simReport}
\title{Generate claim simulation result report in html}
\usage{
simReport(object, simSummary, ...)
}
\arguments{
\item{object}{ClaimType object}

\item{simSummary}{simulation result summary generated by simSummary}

\item{simTriangle}{triangle summary generated by simTriangle}

\item{startDate}{Date after which claims are analyzed}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR}

\item{futureDate}{Date of evaluation for UPR (future claims)}

\item{iYear}{Boolean that indicates whether summary by accident year should be produced in the report}
}
\description{
Generate claim simulation result report in html
}
