% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\docType{methods}
\name{simSummary}
\alias{simSummary}
\alias{simSummary,Simulation,data.frame-method}
\alias{simSummary,ANY-method}
\title{Claim simulation result summary}
\usage{
simSummary(object, simdata, ...)

\S4method{simSummary}{Simulation,data.frame}(object, simdata,
  startDate = as.Date("2012-01-01"),
  evaluationDate = as.Date("2016-12-31"),
  futureDate = as.Date("2017-12-31"))
}
\arguments{
\item{object}{Simulation object}

\item{simdata}{simulation data generated by claimSimulation}

\item{...}{Additional parameters that may or may not be used.}

\item{startDate}{Date after which claims are analyzed;}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR;}

\item{futureDate}{Date of evaluation for UPR (future claims).}
}
\description{
Claim simulation result summary
}
