% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{simSummary}
\alias{simSummary}
\title{Claim simulation result summary}
\usage{
simSummary(object, simdata, ...)
}
\arguments{
\item{object}{Simulation object}

\item{simdata}{simulation data generated by claimSimulation}

\item{startDate}{Date after which claims are analyzed}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR}

\item{futureDate}{Date of evaluation for UPR (future claims)}
}
\description{
Claim simulation result summary
}
