% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\docType{methods}
\name{simTriangle}
\alias{simTriangle}
\alias{simTriangle,Simulation,data.frame,data.frame-method}
\alias{simTriangle,ANY-method}
\title{Claim simulation result triangles}
\usage{
simTriangle(object, claimdata, simdata, ...)

\S4method{simTriangle}{Simulation,data.frame,data.frame}(object, claimdata,
  simdata, frequency = "yearly", startDate = as.Date("2012-01-01"),
  evaluationDate = as.Date("2016-12-31"),
  futureDate = as.Date("2017-12-31"))
}
\arguments{
\item{object}{Simulation object}

\item{claimdata}{claim data used as basis for simulation}

\item{simdata}{simulation data generated by claimSimulation}

\item{...}{Additional parameters that may or may not be used.}

\item{frequency}{triangle frequency, either "yearly" or "quarterly";}

\item{startDate}{Date after which claims are analyzed;}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR;}

\item{futureDate}{Date of evaluation for UPR (future claims).}
}
\description{
Claim simulation result triangles
}
