% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{simTriangle}
\alias{simTriangle}
\title{Claim simulation result triangles}
\usage{
simTriangle(object, claimdata, simdata, ...)
}
\arguments{
\item{object}{Simulation object}

\item{claimdata}{claim data used as basis for simulation}

\item{simdata}{simulation data generated by claimSimulation}

\item{frequency}{triangle frequency, either "yearly" or "quarterly"}

\item{startDate}{Date after which claims are analyzed}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR}

\item{futureDate}{Date of evaluation for UPR (future claims)}
}
\description{
Claim simulation result triangles
}
