% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{methods}
\name{TEKurt}
\alias{TEKurt}
\alias{TEKurt,Normal-method}
\alias{TEKurt,ANY-method}
\alias{TEKurt,Beta-method}
\alias{TEKurt,Exponential-method}
\alias{TEKurt,Gamma-method}
\alias{TEKurt,Geometric-method}
\alias{TEKurt,Lognormal-method}
\alias{TEKurt,NegativeBinomial-method}
\alias{TEKurt,Pareto-method}
\alias{TEKurt,Poisson-method}
\alias{TEKurt,Uniform-method}
\alias{TEKurt,Weibull-method}
\title{Calculate Theoretical Excessive Kurtosis of distribution.
min and max are not applied}
\usage{
TEKurt(object, ...)

\S4method{TEKurt}{Normal}(object)

\S4method{TEKurt}{Beta}(object)

\S4method{TEKurt}{Exponential}(object)

\S4method{TEKurt}{Gamma}(object)

\S4method{TEKurt}{Geometric}(object)

\S4method{TEKurt}{Lognormal}(object)

\S4method{TEKurt}{NegativeBinomial}(object)

\S4method{TEKurt}{Pareto}(object)

\S4method{TEKurt}{Poisson}(object)

\S4method{TEKurt}{Uniform}(object)

\S4method{TEKurt}{Weibull}(object)
}
\arguments{
\item{object}{Distribution Object}

\item{...}{Additional function arguments}
}
\description{
Calculate Theoretical Excessive Kurtosis of distribution.
min and max are not applied
}
\examples{
xPareto <- new("Pareto",p1=20,p2=5)
TEKurt(xPareto)
}
