% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{TEKurt}
\alias{TEKurt}
\title{Calculate Theoretical Excessive Kurtosis of distribution.
min and max are not applied}
\usage{
TEKurt(object, ...)
}
\arguments{
\item{object}{Distribution Object}
}
\description{
Calculate Theoretical Excessive Kurtosis of distribution.
min and max are not applied
}
\examples{
xPareto <- new("Pareto",p1=20,p2=5)
TEKurt(xPareto)
}
