% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{TMean}
\alias{TMean}
\title{Calculate Theoretical Mean of distribution.
min and max are not applied}
\usage{
TMean(object, ...)
}
\arguments{
\item{object}{Distribution Object}
}
\description{
Calculate Theoretical Mean of distribution.
min and max are not applied
}
\examples{
xPareto <- new("Pareto",p1=20,p2=3)
TMean(xPareto)
}
