% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{toString}
\alias{toString}
\title{Object explanation.}
\usage{
toString(object, ...)
}
\arguments{
\item{object}{Object}

\item{...}{Additional function arguments}
}
\description{
Object explanation.
}
\examples{
xPareto <- new("Pareto",p1=20,p2=3)
toString(xPareto)
}
