% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Common.R
\name{truncate}
\alias{truncate}
\title{Truncate a numeric vector}
\usage{
truncate(x, lower, upper)
}
\arguments{
\item{x}{A numeric vector}

\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\description{
Truncate a numeric vector
}
\examples{
trunc(rnorm(100,3,6),0,7)
}
