% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{methods}
\name{TSD}
\alias{TSD}
\alias{TSD,Normal-method}
\alias{TSD,ANY-method}
\alias{TSD,Beta-method}
\alias{TSD,Exponential-method}
\alias{TSD,Gamma-method}
\alias{TSD,Geometric-method}
\alias{TSD,Lognormal-method}
\alias{TSD,NegativeBinomial-method}
\alias{TSD,Pareto-method}
\alias{TSD,Poisson-method}
\alias{TSD,Uniform-method}
\alias{TSD,Weibull-method}
\title{Calculate Theoretical Standard Deviation of distribution.
min and max are not applied}
\usage{
TSD(object, ...)

\S4method{TSD}{Normal}(object)

\S4method{TSD}{Beta}(object)

\S4method{TSD}{Exponential}(object)

\S4method{TSD}{Gamma}(object)

\S4method{TSD}{Geometric}(object)

\S4method{TSD}{Lognormal}(object)

\S4method{TSD}{NegativeBinomial}(object)

\S4method{TSD}{Pareto}(object)

\S4method{TSD}{Poisson}(object)

\S4method{TSD}{Uniform}(object)

\S4method{TSD}{Weibull}(object)
}
\arguments{
\item{object}{Distribution Object}

\item{...}{Additional function arguments}
}
\description{
Calculate Theoretical Standard Deviation of distribution.
min and max are not applied
}
\examples{
xPareto <- new("Pareto",p1=20,p2=3)
TSD(xPareto)
}
