% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{TSkewness}
\alias{TSkewness}
\title{Calculate Theoretical Skewness of distribution.
min and max are not applied}
\usage{
TSkewness(object, ...)
}
\arguments{
\item{object}{Distribution Object}
}
\description{
Calculate Theoretical Skewness of distribution.
min and max are not applied
}
\examples{
xPareto <- new("Pareto",p1=20,p2=4)
TSkewness(xPareto)
}
