% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\docType{methods}
\name{claimSimulation}
\alias{claimSimulation}
\alias{claimSimulation,Simulation-method}
\alias{claimSimulation,ANY-method}
\title{Claim simulation at line/type/status level}
\usage{
claimSimulation(object, ...)

\S4method{claimSimulation}{Simulation}(object, claimData = data.frame(),
  startDate = as.Date("2012-01-01"), evaluationDate = as.Date("2016-12-31"),
  futureDate = as.Date("2017-12-31"), append = TRUE)
}
\arguments{
\item{object}{Simulation object}

\item{...}{Additional parameters that may or may not be used.}

\item{claimData}{claim data including existing claims for RBNER and claim reopenness analysis;}

\item{startDate}{Date after which claims are analyzed;}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR;}

\item{futureDate}{Date of evaluation for UPR (future claims).}

\item{append}{Boolean variable to indicate whether existing simulation results need to be kept.}
}
\description{
Claim simulation at line/type/status level
}
