% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{claimSimulation}
\alias{claimSimulation}
\title{Claim simulation at line/type/status level}
\usage{
claimSimulation(object, ...)
}
\arguments{
\item{object}{Simulation object}

\item{claimData}{claim data including existing claims for RBNER and claim reopenness analysis}

\item{startDate}{Date after which claims are analyzed}

\item{evaluationDate}{Date of evaluation for existing claims and IBNR}

\item{futureDate}{Date of evaluation for UPR (future claims)}
}
\description{
Claim simulation at line/type/status level
}
