% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula.R
\docType{methods}
\name{copulaDataPlot}
\alias{copulaDataPlot}
\alias{copulaDataPlot,CopulaObj-method}
\alias{copulaDataPlot,ANY-method}
\title{Experience data plotting.}
\usage{
copulaDataPlot(object, ...)

\S4method{copulaDataPlot}{CopulaObj}(object)
}
\arguments{
\item{object}{Copula Object}

\item{...}{Additional parameters that may or may not be used}
}
\description{
Experience data plotting.
}
\examples{
library(cascsim)
dist1<-new("Pareto",p1=20,p2=3)
dist2<-new("Normal",p1=5,p2=3,min=0,max=20,truncated=TRUE)
nom.cop <- new("CopulaObj", param=c(0.5),marginal=list(dist1=dist1,dist2=dist2),dimension=2)
setObservation(nom.cop)<-copulaSample(nom.cop,100)
copulaDataPlot(nom.cop)
}
