% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{fitPlot}
\alias{fitPlot}
\title{Compare the raw data and fitted distribution on density, CDF, Q-Q plot and P-P plot}
\usage{
fitPlot(object, ...)
}
\arguments{
\item{object}{FitDist Object}
}
\description{
Compare the raw data and fitted distribution on density, CDF, Q-Q plot and P-P plot
}
\examples{
library(cascsim)
data(claimdata)

#frequecy fitting
findex <- new("Index", startDate = as.Date("2012-01-01"), tabulate=TRUE, monthlyIndex = c(rep(1,11),cumprod(c(1,rep(1.5^(1/12),11))),cumprod(c(1.5,rep((1.3/1.5)^(1/12),11))),cumprod(c(1.3,rep((1.35/1.3)^(1/12),11))),cumprod(c(1.35,rep((1.4/1.35)^(1/12),11))),1.4))
rawdata <- as.data.frame(as.Date(claimdata[(claimdata[,"LoB"]=="Auto" & claimdata[,"Type"]=="H"),]$OccurDate))
colnames(rawdata)<-"OccurDate"
xFit <- new("FitDist", observation=rawdata, trend=findex,startDate = as.Date("2012-01-01"),method="mle",ifreq=TRUE,idate=TRUE, freq="Monthly")
xFit <- setFitdata(xFit)
setTrialDist(xFit) <- new("Poisson")
xFit@soutput
fitPlot(xFit)
}
