% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{PDFPlot}
\alias{PDFPlot}
\title{Plotting the PDF of data and fitted distribution}
\usage{
PDFPlot(object, ...)
}
\arguments{
\item{object}{FitDist Object}

\item{...}{Additional function arguments}
}
\description{
Plotting the PDF of data and fitted distribution
}
