% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula.R
\docType{methods}
\name{setCopulaType<-}
\alias{setCopulaType<-}
\alias{setCopulaType<-,CopulaObj,character-method}
\alias{setCopulaType,ANY-method}
\title{Set copula type.}
\usage{
setCopulaType(this, ...) <- value

\S4method{setCopulaType}{CopulaObj,character}(this) <- value
}
\arguments{
\item{this}{Copula Object}

\item{...}{Additional function arguments}

\item{value}{The copula type}
}
\description{
Set copula type.
}
\examples{
library(cascsim)
dist1<-new("Pareto",p1=20,p2=3)
dist2<-new("Normal",p1=5,p2=3,min=0,max=20,truncated=TRUE)
cop <- new("CopulaObj", param=c(0.5),marginal=list(dist1=dist1,dist2=dist2),dimension=2)
setCopulaType(cop) <- "joe"
}
