% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFac.R
\docType{methods}
\name{setFun<-}
\alias{setFun<-}
\alias{setFun<-,DevFac,character-method}
\alias{setFun,ANY-method}
\title{Set the model format/link function (identity/inverse/log/exponential).
Only used when FacModel == TRUE.}
\usage{
setFun(this, ...) <- value

\S4method{setFun}{DevFac,character}(this) <- value
}
\arguments{
\item{this}{DevFac Object}

\item{...}{Additional function arguments}

\item{value}{String Value (default:"identity")}
}
\description{
Set the model format/link function (identity/inverse/log/exponential).
Only used when FacModel == TRUE.
}
\examples{
xIBNERFactor <- new("DevFac")
setID(xIBNERFactor)<-"IF1"
setFacModel(xIBNERFactor)<-TRUE
setFun(xIBNERFactor)<-"identity"
setXname(xIBNERFactor)<- c("x1","x2","x3")
setParas(xIBNERFactor)<-c(0.6,-0.2,0.01,-0.3,0.02,0.03,0.01,0.02)
xIBNERFactor<-setDevFac(xIBNERFactor)
xIBNERFactor
}
