% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DevFac.R
\docType{methods}
\name{setMeanList<-}
\alias{setMeanList<-}
\alias{setMeanList<-,DevFac,vector-method}
\alias{setMeanList,ANY-method}
\title{Set the year-to-year loss development factor.}
\usage{
setMeanList(this, ...) <- value

\S4method{setMeanList}{DevFac,vector}(this) <- value
}
\arguments{
\item{this}{DevFac Object}

\item{...}{Additional function arguments}

\item{value}{Numeric Vector}
}
\description{
\code{setMeanList<-} sets expected year-to-year loss development factor. Years after It is only used when ibnerfModel == FALSE.
}
\examples{
xIBNERFactor <- new("DevFac")
setID(xIBNERFactor)<-"IF1"
setFacModel(xIBNERFactor)<-FALSE
setMeanList(xIBNERFactor)<-c(1.26,1.1,1.05,1.02,1)
setVolList(xIBNERFactor)<-rep(0.02,5)
xIBNERFactor
}
