% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIndex.R
\docType{methods}
\name{setMonthlyIndex<-}
\alias{setMonthlyIndex<-}
\alias{setMonthlyIndex<-,Index,vector-method}
\alias{setMonthlyIndex,ANY-method}
\title{Set monthly index values.}
\usage{
setMonthlyIndex(this, ...) <- value

\S4method{setMonthlyIndex}{Index,vector}(this) <- value
}
\arguments{
\item{this}{Index Object}

\item{...}{Additional function arguments}

\item{value}{Numeric Vector}
}
\description{
\code{setMonthlyIndex<-} sets monthly index values.
}
\examples{
xindex <- new("Index")
setID(xindex)<-"IDX1"
setTabulate(xindex)<-TRUE
setMonthlyIndex(xindex)<- rep(1,360)
xindex<-setIndex(xindex)
xindex@monthlyIndex
}
