% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDist.R
\name{setTrialDist<-}
\alias{setTrialDist<-}
\title{Distribution fitting and testing.}
\usage{
setTrialDist(this) <- value
}
\arguments{
\item{this}{FitDist Object}

\item{value}{Distribution to fit to}
}
\description{
Distribution fitting and testing.
}
\examples{
library(cascsim)
data(claimdata)

#frequecy fitting example 1
findex <- new("Index", startDate = as.Date("2012-01-01"), tabulate=TRUE, monthlyIndex = c(rep(1,11),cumprod(c(1,rep(1.5^(1/12),11))),cumprod(c(1.5,rep((1.3/1.5)^(1/12),11))),cumprod(c(1.3,rep((1.35/1.3)^(1/12),11))),cumprod(c(1.35,rep((1.4/1.35)^(1/12),11))),1.4))
rawdata <- as.data.frame(as.Date(claimdata[(claimdata[,"LoB"]=="Auto" & claimdata[,"Type"]=="H"),]$occurrenceDate))
colnames(rawdata)<-"occurrenceDate"
xFit <- new("FitDist", observation=rawdata, trend=findex,startDate = as.Date("2012-01-01"),method="mle",ifreq=TRUE,idate=TRUE, freq="Monthly")
xFit <- setFitdata(xFit)
setTrialDist(xFit) <- new("Poisson")
xFit@soutput
observationPlot(xFit)
fitPlot(xFit)

#frequecy fitting example 2            
rawdata <- as.data.frame(as.Date(claimdata[(claimdata[,"LoB"]=="Liab" & claimdata[,"Type"]=="N"),]$occurrenceDate))
colnames(rawdata)<-"occurrenceDate"
xFit <- new("FitDist", observation=rawdata, trend=findex,startDate = as.Date("2012-01-01"),method="qme",ifreq=TRUE,idate=TRUE, freq="Monthly")
xFit@probs <- c(0.5,0.9)
xFit <- setFitdata(xFit)
setTrialDist(xFit) <- new("NegativeBinomial",p1=20,p2=0.5)
xFit@soutput
observationPlot(xFit)
fitPlot(xFit)

#severity fitting example 1
sindex <- new("Index", startDate = as.Date("2012-01-01"), monthlyIndex = cumprod(c(1,rep(1.03^(1/12),59))))
sdate <- as.Date("2012-01-01")
rawdata <- as.data.frame(claimdata[(claimdata[,"LoB"]=="Liab" & claimdata[,"Type"]=="H" & claimdata[,"status"]=="Closed" & as.Date(claimdata[,"occurrenceDate"])>=sdate),][,c(9,7)])
xFit <- new("FitDist", observation=rawdata, trend=sindex,startDate = sdate,method="qme",probs=c(0.7,0.5),ifreq=FALSE)
#xFit <- new("FitDist", observation=rawdata, trend=sindex,startDate = sdate,method="mme",ifreq=FALSE)
xFit <- setFitdata(xFit)
setTrialDist(xFit) <- new("Pareto",p1=500,p2=2.5)
xFit@soutput
observationPlot(xFit)
fitPlot(xFit)

#severity fitting example 2 Creating empirical distribution
sindex <- new("Index", startDate = as.Date("2012-01-01"), monthlyIndex = cumprod(c(1,rep(1.03^(1/12),59))))
sdate <- as.Date("2012-01-01")
rawdata <- as.data.frame(claimdata[(claimdata[,"LoB"]=="Liab" & claimdata[,"Type"]=="H" & claimdata[,"status"]=="Closed" & as.Date(claimdata[,"occurrenceDate"])>=sdate),][,c(9,7)])
xFit <- new("FitDist", observation=rawdata, trend=sindex, ifreq=FALSE)
xFit <- setFitdata(xFit)
setTrialDist(xFit) <- new("Empirical")
xFit@soutput
xFit@fitted@empirical
observationPlot(xFit)
fitPlot(xFit)

#severity fitting example 3
sindex <- new("Index", startDate = as.Date("2012-01-01"), monthlyIndex = cumprod(c(1,rep(1.03^(1/12),59))))
sdate <- as.Date("2012-01-01")
rawdata <- as.data.frame(claimdata[(claimdata[,"LoB"]=="Auto" & claimdata[,"Type"]=="H" & claimdata[,"status"]=="Closed" & as.Date(claimdata[,"occurrenceDate"])>=sdate),][,c(9,7)])
xFit <- new("FitDist", observation=rawdata, trend=sindex,startDate = sdate,method="mle",ifreq=FALSE)
xFit <- setFitdata(xFit)
setTrialDist(xFit) <- new("Lognormal",p1=2,p2=1,truncated=TRUE,min=0.1)
xFit@soutput
observationPlot(xFit)
fitPlot(xFit)

#report lag fitting
lindex <- new("Index", startDate = as.Date("2012-01-01"), monthlyIndex = rep(1,360))
rawdata <- as.data.frame(claimdata[(claimdata[,"LoB"]=="Auto" & claimdata[,"Type"]=="H"),][,c(5,6)])
rawdata <- as.numeric(as.Date(rawdata$reportDate)-as.Date(rawdata$occurrenceDate))
rawdata <- as.data.frame(rawdata)
xFit <- new("FitDist", observation=rawdata, trend=lindex,startDate = as.Date("2012-01-01"),method="mle",ifreq=FALSE)
xFit <- setFitdata(xFit)
setTrialDist(xFit) <- new("Exponential",p1=1)
xFit@soutput
fitPlot(xFit)
observationPlot(xFit)
observationPlot(xFit)
fitPlot(xFit)

#payment (after deductible and limit) data fitting (truncated distribution) 
sindex <- new("Index", startDate = as.Date("2012-01-01"), monthlyIndex = cumprod(c(1,rep(1^(1/12),59))))
rawdata <- data.frame(rtlnorm(10000,3,2,5,100))
xFit <- new("FitDist", observation=rawdata, trend=sindex,startDate = as.Date("2012-01-01"),method="mle",ifreq=FALSE)
xFit <- setFitdata(xFit)
setTrialDist(xFit) <- new("Lognormal",p1=2,p2=1,min=5,max=100,truncated=TRUE)
xFit@soutput
observationPlot(xFit)
fitPlot(xFit)
xFit <- new("FitDist", observation=rawdata, trend=sindex,startDate = as.Date("2012-01-01"),method="qme",probs=c(0.3,0.7),ifreq=FALSE)
xFit <- setFitdata(xFit)
setTrialDist(xFit) <- new("Lognormal",p1=2,p2=1,min=5,max=100,truncated=TRUE)
xFit@soutput
observationPlot(xFit)
fitPlot(xFit)
}
